requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
google-auth==2.33.0
